+++ src/varintExternalBigEndian.h
@@ -57,10 +57,10 @@ uint64_t varintExternalBigEndianGet(const void *p, varintWidth encoding);
             (result) = (_vimp_src)[0];                                         \
             break;                                                             \
         case VARINT_WIDTH_16B:                                                 \
-            (result) = _vimp_src[0] << 8 | _vimp_src[1];                       \
+            (result) = (uint64_t)(_vimp_src[0]) << 8 | (uint64_t)(_vimp_src[1]); \
             break;                                                             \
         case VARINT_WIDTH_24B:                                                 \
-            (result) = _vimp_src[0] << 16 | _vimp_src[1] << 8 | _vimp_src[2];  \
+            (result) = (uint64_t)(_vimp_src[0]) << 16 | (uint64_t)(_vimp_src[1]) << 8 | (uint64_t)(_vimp_src[2]); \
             break;                                                             \
         default:                                                               \
             (result) = varintExternalBigEndianGet((src), (width));             \
